sensors/clustering/requirements.txt
matplotlib==3.1.3
pandas
numpy
statsmodels
scikit-learn
svgpath2mpl
colour
pyyaml